moved wordsFrench, wordsGerman, words,Spainish into the test data folder, I have also updated the model and contoller for words and passed their tests again
